+++ src/main/java/roomescape/exception/GlobalExceptionHandler.java
@@ -1,0 +1,20 @@
+package roomescape.exception;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(HttpMessageNotReadableException.class)
+    private ResponseEntity<String> handleJsonParseError(HttpMessageNotReadableException e) {
+        return ResponseEntity.badRequest().body("요청 형식이 올바르지 않습니다.");
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    private ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+}
